requirements.txt
pygame>=2.0.0
numpy>=1.21.0
supabase>=2.0.0
python-dotenv>=1.0.0
